% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba-package.R
\docType{data}
\name{reddit}
\alias{reddit}
\title{Example data from reddit}
\format{An object of class \code{list} of length 2.}
\source{
\url{https://www.reddit.com/dev/api/}
}
\usage{
reddit
}
\description{
Example data from reddit
}
\keyword{datasets}
