% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULLs}
\usage{
replace_null(x, replacement = NA)
}
\arguments{
\item{x}{list to use}

\item{replacement}{Replacement value for missing values}
}
\description{
Replace all the empty values in a list
}
\examples{
list(a = NULL, b = 1, c = list(foo = NULL, bar = NULL)) \%>\% replace_null()

}
