% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba-package.R
\docType{package}
\name{roomba-package}
\alias{roomba-package}
\title{roomba package}
\description{
\tabular{ll}{
Package: \tab roomba\cr
Type: \tab Package\cr
}
}
\details{
Collection of functions to do deal with deeply nested data
}
