% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrroomba.R
\name{roomba}
\alias{roomba}
\title{Roomba}
\usage{
roomba(inp, replacement = NA, cols = NULL, keep = all)
}
\arguments{
\item{inp}{input list}

\item{replacement}{Replacement for NULL values. Defaults to NA_character_.}

\item{cols}{column names to extract}

\item{keep}{Select \code{all} or \code{any} columns.}
}
\description{
Tidy your nested list
}
