% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{shiny_roomba}
\alias{shiny_roomba}
\title{Run the roomba app}
\usage{
shiny_roomba(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Run the roomba app
}
\examples{
\dontrun{
if(require(shiny)){
   shiny_roomba()
}
}
}
