% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba-package.R
\docType{data}
\name{simple}
\alias{simple}
\title{Example data with simple nested list}
\format{An object of class \code{character} of length 1.}
\usage{
simple
}
\description{
Example data with simple nested list
}
\keyword{datasets}
