% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba-package.R
\docType{data}
\name{twitter_data}
\alias{twitter_data}
\title{Example Twitter data}
\format{An object of class \code{list} of length 20.}
\source{
\url{https://developer.twitter.com/en/docs}
}
\usage{
twitter_data
}
\description{
Example data from Twitter API
}
\details{
twitter_data[[1]][["id"]]

[[1]: R:[1
["id"]: R:%22id%22
}
\keyword{datasets}
