% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba-package.R
\docType{data}
\name{vimeo}
\alias{vimeo}
\title{Example data from vimeo}
\format{An object of class \code{response} of length 10.}
\source{
\url{https://developer.vimeo.com/}
}
\usage{
vimeo
}
\description{
Example data from vimeo
}
\keyword{datasets}
