% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomba.R
\name{roomba}
\alias{roomba}
\title{Roomba}
\usage{
roomba(inp, replacement = NA_character_, rows = NULL, cols = NULL,
  clean = TRUE)
}
\arguments{
\item{inp}{Element to replace}

\item{replacement}{Replacement for NULL values. Defaults to NA_character_.}

\item{rows}{rows to iterate over}

\item{cols}{columns to include}

\item{clean}{Should NULL values be changed to NAs?}
}
\description{
Tidy your nested list
}
