% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrroomba.R
\name{roomba}
\alias{roomba}
\title{Roomba}
\usage{
roomba(inp, clean = TRUE, replacement = NA_character_, rows = NULL,
  cols = NULL)
}
\arguments{
\item{inp}{Element to replace}

\item{clean}{Should NULL values be changed to NAs?}

\item{replacement}{Replacement for NULL values. Defaults to NA_character_.}

\item{rows}{id to label rows}

\item{cols}{columns to include}
}
\description{
Tidy your nested list
}
